Comments and GO analysis
Added more comments and commented out several lines of code that are useful if running the first iteration of the FASTQ files. Add code to assist in the GO term analysis; however, the reference genome we need (mm10) does not contain all of the necessary requirements for a completed analysis.